Extract useAssistant.
